Refactor FindPrecedingHeader[Indent]
- change the name to better reflect the purpose;
- add line argument to avoid relying on the cursor position;
- return only indent, line number isn't used and can be incorrect;
- properly handle clause headers and suites on a single line.

Rename minindent to maxindent as it is used as the maximum indent.

This function terminates when the maxindent is 0 to avoid processing
top level strings, but the returned line number isn't the header line
in this case.

Process lines with indent equal to maxindent to properly handle cases
where header and suite are on the same line. Example:
if (i > 0): print('+')
elif (i < 0): print('-')
else: print('0')